==================POST 2.1==================
This patch is a total re-working of the way that ntop detects openSSL.
It now does a test compile, using the ERR_error_string_n() function
which is required, but was 1st introduced in 0.9.6...

It was excluded from 2.1 due to an insufficient ability to test in time.

There is a new ./configure flag, --enable-showopenssltests, which
will show the details of the openssl tests in the output.

I also moved the handling of --enable-sslwatchdog so that it aborts if
you don't have openssl.

I choose to make --enable-sslv3 just a warning if you don't have openssl.
I felt it was just less severe.

This also tests for (future) versions of openssl (0.9.7) with a warning
that they didn't exist or were beta and thus untested at the time of
ntop 2.1's release.

-----Burton

   ==================POST 2.1==================CVS: ----------------------------------------------------------------------