Update to ntop's mtu and headerSize tables.
This won't fix taviso's PPPoA problem, it's the prep for it.

ntop now has a table that can handle what's currently defined in the cvs
at http://www.tcpdump.org for bpf.h's DLT_ constants.  And a mechanism for
responding to changes in that table.

This fix adds reporting and warning messages so that we can detect and
improve the actual handling of these situations.

The one processing issue - in nfsPlugin.c - has been corrected.

Note that invalid/unknown data in the tables does not cause ntop to fail,
only to report erroneous classification of packets WRT mtu values.

Also note that you, the ISP and everybody in between has always had the ability
to change the mtu values.  All we can do is asume, say, Ethernet is 1500 bytes +
14 byte header.  If the ISP has an 800 byte link, well, the real mtu is 800, but
ntop will use the 1514 value.  Why?  Because determining the actual mtu is a black
art.  I've seen it done by starting at 1514 and sending packets, each one smaller,
until it doesn't come back with fragmentation required.  Ugly Ugly Ugly...

-----Burton (ref 127)