This is a configure.in / configure fix to handle a bug in automake 1.7.1
ntop distributes configure (and other files) created using autoconf 2.53
and automake 1.6.3.  These files work with automake 1.6.x and 1.7.
But not automake 1.7.1.  (automake 1.5 is unknown, but suspected NOT to work).

It's a deliberate change on the part of automake, that breaks ANY chance
of distributing ONE set of files which can be used by all ntop users -
whatever version of the auto* tools they have...  <GRRR />

Reference: http://mail.gnu.org/pipermail/automake-patches/2002-January/000559.html

I can not trap it cleanly, because the call to automake which reports the problem
happens before the FIRST generated line in ./configure.

We can't just fix up the one file - they all have to be build by a consistent set
of tools, and if there are problems, then the "make" step fails.

What this patch does is to detect the problem, perform the best mitigation
possible and then abort ./configure.  This happens well after the automake
error messages, but there is a pretty clear error box displayed.

If the auto* tools execute cleanly (thus creating a 1.7.1 set in the
ntop source directory), the user is instructed to rerun ./configure,
"with whatever parameters you normally use plus --enable-automake171".

-----Burton (ref 151)