Add a "HELP" page to ntop's web interface.  It's off the About frame (i.e. seen first on startup) and it has pointers to the mailing lists, sourceforge, etc.
There is also a conversion of the FAQ to html (awk script, not pretty, not perfect
but far better than nothing).

I'm not sure that the conversion will be done automatically, so if you don't see the
page, you need to do a "make faq.html" -- just like you need to (manually) do
"make ntop.txt" and/or "make ntop.html" after the man page (ntop.8) gets updated.
But I *think* this will happen automatically.

-----Burton (ref 165)