Hosts that speak over the JetDirect port are marked as printers.
Added the concept of host contacts. Basically this is an index that contains the number of hosts contacted by each host. This is a value that shows how many different hosts have been contacted. Hosts that run a DNS or mail server should have high value for this index. Workstations with a high contact index may indicate that either the host is scanning (is scanned) the network or that it has a trojan/virus.