Bleep it.
I've removed the --enable-automake171 flag in favor of an explicit test for
matching auto* tools versions.  That is, there is a file distributed (new)
with ntop - autotoolversions - that has the exact version of the three
auto* tools (automake, autoconf and libtool) used to create the versions
of the various generated files (configure, Makefile, etc.) which are
distributed with ntop.

If those versions do not match - EXACTLY - then the first time you run
./configure, ntop will delete and recreate ALL of the generated files
and update autotoolversions.

You will then need to run ./configure a SECOND time to do the actual
configuration!

But, this should eliminate the problems with conflicting versions.

It does not change the minimum required versions:

    automake 1.6+
    autoconf 2.5x+
    libtool 1.3+ (we say 1.4+, but it does seem to work just fine with 1.3.4)

-----Burton (ref 188)