BUGFIX:
The -s | --no-promisc mode is documented such that users might believe
they don't have to be root to run ntop.  And in fact, on some systems
that might work.  This patch makes the code behave as documented.

(Basically it will try and do the pcap_open - if it fails, you get a
sorry message.  If it works, ntop continues) (vs. bouncing you way
earlier before even trying).

-----Burton (ref 260)