1. Add POSIX _atfork() handler to cleanup mutexes after a fork().
2. Reduce failed http request alarm timeout to 15 seconds from 2 minutes
   and add a log message reporting the failure.

This patch also fixes/bypasses a known problem with glibc 2.3.x, which
manifests itself as random crashes, usually of the ntop web server.
It appears - although not well understood - that a free() of memory in a
child which was malloc()ed in the parent before the fork() call - i.e.
where the child has a completly separate copy of the entire memory
space - somehow corrupts the malloc() areas in the PARENT.  This is
supposed to be impossible, but combining fork() and POSIX pthread is
an edge case anyway...

So...

3. Add childntoppid to myGlobals, to allow us to detect whether we're
   in a fork()ed child or a parent/POSIX thread.

and

4. A known problem spot (util.c around line 3370) now avoids doing the
   free.  Since we're updating the OS information, only to throw it all away
   when the fork()ed child exits (remember this is a separate copy of memory,
   not the real myGlobals), it seems reasonable to avoid the whole block of
   code.


If random crashes persist, there is commented out code in leaks.c
(search for DIAGNOSTIC) to issue log messages in cases where we free()
memory in a fork()ed child.  These COULD be problems, if the malloc() was
in the parent.  Because there's no way to differentiate between the two
cases, the message will produce a lot of noise.  If you experience the
random crashes, try enabling it for a little while and report the last few
messages before it bombs...

-----Burton (ref 283, originally developed for ntop_2_2_patches)