Add --disable-instantsessionpurge run time option.
Due to a bug, ntop reclassifies a properly closed session to FLAG_STATE_TIMEOUT.

FLAG_STATE_TIMEOUT is used in the tcp/ip stack to represent a session which is
initiated but never completes the 3-way handshake.  These are symptomatic of a
busted network, or most commonly a DOS attack against a web server.  Because
there's little information value vs. the huge # of sessions, FLAG_STATE_TIMEOUT
sessions are purged almost instantly (instead of using the session timeout
value).

So, at present, ntop doesn't retain session data for more than a few seconds
after the tcp/ip session is terminated.

If you enable the --disable-instantsessionpurge switch, the session state
stays FLAG_STATE_FIN2_ACK2 or FLAG_FLAG_STATE_SYN_ACK, which allows ntop
to report on the session until the (default 5 minute) timeout interval
expires.

This correct behavior is NOT the default, because a large number of retained
sessions can significantly increase the memory usage of the ntop instance if
there are hosts that open/close large #s of sessions (e.g. a web server).

Look for this to flip/flop to an --enable-instantsessionpurge in 2.3.1
(development versions for 2.4).

Also, fixed up the sessions report to correctly show Duration (it's last
seen - first seen) and added an 'Inactive' column.  This affects all
instances, regardless of the run time option setting.


-----Burton (ref 282)