Put the pcap_stats back into the PR report.
Luca says it blows up on Win32 if all the plugins are stopped.
On the face of it's that's silly - the plugins don't reference this
chunk of memory.  But, he know what he's been seeing.

So while it's something else under the covers and I've got no clue why,
but this is often valuable information, I put it back it, just #ifndef WIN32...

-----Burton (ref 338)