BMS0407
 see plugins/xmldumpPlugin.c for this commit.

Note that the setting of the (strange) places for the required libraries is now
in the configureextra files, and only LINUXredhat9 is actually populated.

This is a new way for telling ntop about files located in strange places on
specific OSes or distributions, but it makes the ./configure process very clean
for OSes/distros that have standard pre-built libraries (xBSD ports, RPMs, .debs,
etc. - also Gentoo ebuilds).

-----Burton