BMS0434
Add 'final' hooks in ./configure.

This invokes - if they exist - configureextra/final_xxxxx just like the
preliminary configureextra files.

This occurs after all of the tests and this line:

Removing dups and misplaced entries from LIBS and INCS...

and before we write out the configuration and create Makefile:

===================================================================

This is your ntop 2.2.99 configuration:
...

So, theoretically you can use these for any last minute cleanup. Be aware that
ntop has already run all of it's tests, so it's quite possible to break things.

In fact, there are no legit uses of this I can think of at present, but there's
one user who wants it.  And it's symetrical... s'be'it...

-----Burton