"Fix" for the 'RedHat 8 hangs' problem.
In my testing for this problem, I have seen all of ntop's threads hanging
in the syslog() routines.

This "fix" adds a mutex around the calls.  It fixes the problem for me, at
a cost of some slight single threading.  Since the problem shows up most
often at startup when a bunch of threads are issuing their initial messages,
that's not a big hit.  Still, I can't PROVE it's the fix, so it's
conditional on a -D defined item (configureextra/LINUXredhat8.0:

     CPPFLAGS="${CPPFLAGS} -DFORPRENPTL"

)

Here's where it gets odd...

If you skulk around the web, you'll see a lot of threads that talk about random
hangs in syslog() routines, across lots of different open source projects.
It appears that syslog() routines aren't thread safe... and the usual 'answer'
is to create some sort of thread-safe wrapper.  That's what I've done here.

Yet, strangely, I've never seen this w/ ntop before 3.1.

If this is really the problem, then other pre-NPTL Linuxes could also have
the same hang.  If so, we will need to adjust their configureextra/LINUXxxxx
routines like the change this fix makes in configureextra/LINUX/redhat8.0!

-----Burton (ref 541)