Add packet delay timing report to info.html.
It's a moving average of last 1K (adjustable) packets, showing time
between libpcap time stamp (receipt) and start of processPacket().

The calculation in the note gives a VERY ROUGH approximation of the
rate of packet arrival ntop can handle.

In globals-defines.h, #undef MAX_ARRIVAL_BUFFER to disable this feature!

-----Burton (ref 571)